Add dry run flag

+++ helpers/files.rb
@@ -1,5 +1,6 @@
 require 'fileutils'
 require './helpers/shared'
+
 def ignored_containers
   ignore_file_path = File.join(Dir.getwd, '.dbignore')
 
@@ -11,17 +12,19 @@ end
 def work_folder
   path = File.join(Dir.getwd, 'tmp', run_date)
 
-  FileUtils.mkdir_p path unless File.directory? path
+  FileUtils.mkdir_p path unless File.directory? path || $dry_run
 
   path
 end
 
 def clean_work_folder
-  FileUtils.remove_dir work_folder
+  FileUtils.remove_dir work_folder unless $dry_run
 end
 
 def zip_files(paths, output)
   out_path = File.join(work_folder, output)
-  `zip -uj #{out_path} #{paths.join(' ')}`
+  
+  run_cmd "zip -uj #{out_path} #{paths.join(' ')}"
+
   out_path
 end
